+++ src/storage/storage_adapter/zip.rs
@@ -2,10 +2,10 @@
 
 use crate::{
     array::codec::extract_byte_ranges_read,
-    byte_range::ByteRange,
     storage::{
-        storage_value_io::StorageValueIO, Bytes, ListableStorageTraits, ReadableStorageTraits,
-        StorageError, StoreKey, StoreKeys, StoreKeysPrefixes, StorePrefix, StorePrefixes,
+        byte_range::ByteRange, storage_value_io::StorageValueIO, Bytes, ListableStorageTraits,
+        ReadableStorageTraits, StorageError, StoreKey, StoreKeys, StoreKeysPrefixes, StorePrefix,
+        StorePrefixes,
     },
 };