+++ src/storage/store/store_async/object_store.rs
@@ -1,13 +1,10 @@
 use futures::{StreamExt, TryStreamExt};
 use object_store::path::Path;
 
-use crate::{
-    byte_range::ByteRange,
-    storage::{
-        AsyncBytes, AsyncListableStorageTraits, AsyncReadableStorageTraits,
-        AsyncReadableWritableStorageTraits, AsyncWritableStorageTraits, MaybeAsyncBytes,
-        StorageError, StoreKey, StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix,
-    },
+use crate::storage::{
+    byte_range::ByteRange, AsyncBytes, AsyncListableStorageTraits, AsyncReadableStorageTraits,
+    AsyncReadableWritableStorageTraits, AsyncWritableStorageTraits, MaybeAsyncBytes, StorageError,
+    StoreKey, StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix,
 };
 
 impl From<object_store::Error> for StorageError {